pyproject.toml
[tool.poetry]
name = "relevance-pipeline"
version = "0.0.1"
description = "relevance pipeline is a tool to extract, store and retrieve relevant information from a set of documents."
authors = ["Alex Jenkins <alex.jenkins@ubank.com.au>"]
readme = "README.md"
packages = [{include = "relevance_pipeline", from = "."}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
python-dotenv = "^1.0.0"
chromadb = "^0.5.0"
llama-index = "^0.10.39"
deepeval = "^0.21.45"
llama-index-readers-web = "^0.1.17"
pymupdf = "^1.24.4"
llama-index-retrievers-bm25 = "^0.1.3"
llama-index-vector-stores-qdrant = "^0.2.8"


[tool.poetry.group.dev.dependencies]
bumpversion = "^0.6.0"
pre-commit = "^3.5.0"
pytest = "^8.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gentlemen = { callable = "relevance_pipeline.__main__:main"}
